+++ lib/reports/ReportQuery.ts
@@ -14,8 +14,10 @@ export interface Query {
 	from: ISODate;
 	/** End of query time interval. */
 	to: ISODate;
-	/** Optional, list of Toggl project IDs to include/exclude. */
+	/** Optional, list of Toggl project IDs or names to include/exclude. */
 	projectSelection?: Selection;
+	/** Optional, list of Toggl client IDs or names to include/exclude. */
+	clientSelection?: Selection;
 	/** Optional, indicates a sort order for rendered results. */
 	sort?: SortOrder;
 	/** Optional, indicates a grouping for list reports. */
@@ -56,8 +58,10 @@ export enum SortOrder {
 }
 
 export enum GroupBy {
-	/** Group list of time entries by project. */
+	/** Group list of time entries by client. */
 	PROJECT = 'PROJECT',
+	/** Group list of time entries by project. */
+	CLIENT = 'CLIENT',
 	/** Group list of time entries by date. */
 	DATE = 'DATE'
 }