+++ lib/reports/parser/parseSelection.spec.ts
@@ -17,8 +17,10 @@ describe('parseSelection', () => {
 		testParseSelection({
 			input: [Keyword.INCLUDE, Keyword.PROJECTS, 'project A', 1234567890],
 			query: test_query,
-			mode: SelectionMode.INCLUDE,
-			list: ['project A', 1234567890],
+			projects: {
+				mode: SelectionMode.INCLUDE,
+				list: ['project A', 1234567890]
+			},
 			remaining: []
 		});
 	});
@@ -27,8 +29,62 @@ describe('parseSelection', () => {
 		testParseSelection({
 			input: [Keyword.EXCLUDE, Keyword.PROJECTS, 'project A', 1234567890],
 			query: test_query,
-			mode: SelectionMode.EXCLUDE,
-			list: ['project A', 1234567890],
+			projects: {
+				mode: SelectionMode.EXCLUDE,
+				list: ['project A', 1234567890]
+			},
+			remaining: []
+		});
+	});
+
+	test('"INCLUDE CLIENTS" adds clientSelection to query object', () => {
+		testParseSelection({
+			input: [Keyword.INCLUDE, Keyword.CLIENTS, 'client A', 'client B'],
+			query: test_query,
+			clients: {
+				mode: SelectionMode.INCLUDE,
+				list: ['client A', 'client B']
+			},
+			remaining: []
+		});
+	});
+
+	test('"EXCLUDE CLIENTS" adds clientSelection to query object', () => {
+		testParseSelection({
+			input: [Keyword.EXCLUDE, Keyword.CLIENTS, 'client A', 'client B'],
+			query: test_query,
+			clients: {
+				mode: SelectionMode.EXCLUDE,
+				list: ['client A', 'client B']
+			},
+			remaining: []
+		});
+	});
+
+	test('Adding project selection after client selection', () => {
+		test_query.clientSelection = {
+			mode: SelectionMode.INCLUDE,
+			list: ['client A']
+		};
+		testParseSelection({
+			input: [Keyword.EXCLUDE, Keyword.PROJECTS, 'project B'],
+			query: test_query,
+			clients: { mode: SelectionMode.INCLUDE, list: ['client A'] },
+			projects: { mode: SelectionMode.EXCLUDE, list: ['project B'] },
+			remaining: []
+		});
+	});
+
+	test('Adding client selection after project selection', () => {
+		test_query.projectSelection = {
+			mode: SelectionMode.INCLUDE,
+			list: ['project A']
+		};
+		testParseSelection({
+			input: [Keyword.EXCLUDE, Keyword.CLIENTS, 'client B'],
+			query: test_query,
+			clients: { mode: SelectionMode.EXCLUDE, list: ['client B'] },
+			projects: { mode: SelectionMode.INCLUDE, list: ['project A'] },
 			remaining: []
 		});
 	});
@@ -38,8 +94,6 @@ describe('parseSelection', () => {
 			{
 				input: [Keyword.EXCLUDE, Keyword.PROJECTS, Keyword.INCLUDE],
 				query: test_query,
-				mode: null,
-				list: null,
 				remaining: null
 			},
 			/must be followed by at least one item/g
@@ -51,8 +105,6 @@ describe('parseSelection', () => {
 			{
 				input: [Keyword.WEEKS, Keyword.PROJECTS, 'project A', 1234567890],
 				query: test_query,
-				mode: null,
-				list: null,
 				remaining: null
 			},
 			/Invalid token/g
@@ -64,11 +116,44 @@ describe('parseSelection', () => {
 			{
 				input: [Keyword.EXCLUDE, Keyword.WEEKS, 'project A', 1234567890],
 				query: test_query,
-				mode: null,
-				list: null,
 				remaining: null
 			},
 			/Invalid token/g
+		);
+	});
+
+	it('fails on pre-existing project selection', () => {
+		test_query.projectSelection = { mode: SelectionMode.EXCLUDE, list: [] };
+		testParseSelection(
+			{
+				input: [Keyword.EXCLUDE, Keyword.PROJECTS, 'Abcd', 1234567890],
+				query: test_query,
+				remaining: null
+			},
+			/A query can only contain a single selection expression for keyword/g
+		);
+	});
+
+	it('fails on pre-existing client selection', () => {
+		test_query.clientSelection = { mode: SelectionMode.EXCLUDE, list: [] };
+		testParseSelection(
+			{
+				input: [Keyword.EXCLUDE, Keyword.CLIENTS, 'Abcd', 1234567890],
+				query: test_query,
+				remaining: null
+			},
+			/A query can only contain a single selection expression for keyword/g
+		);
+	});
+
+	it('fails on client filtering using numeric ID', () => {
+		testParseSelection(
+			{
+				input: [Keyword.EXCLUDE, Keyword.CLIENTS, 'Abcd', 1234567890],
+				query: test_query,
+				remaining: null
+			},
+			/Filtering by numeric client id is not currently supported/g
 		);
 	});
 });
@@ -76,9 +161,15 @@ describe('parseSelection', () => {
 interface SelectionTestParams {
 	input: Token[];
 	query: Query;
-	mode: SelectionMode;
-	list: (string | number)[];
 	remaining: Token[];
+	projects?: {
+		mode: SelectionMode;
+		list: (string | number)[];
+	};
+	clients?: {
+		mode: SelectionMode;
+		list: (string | number)[];
+	};
 }
 
 function testParseSelection(
@@ -89,12 +180,20 @@ function testParseSelection(
 
 	if (expectedError == undefined) {
 		expect(parser.parse(params.input, params.query)).toEqual(params.remaining);
-
-		expect(params.query.projectSelection).toEqual(expect.anything());
-		expect(params.query.projectSelection).toMatchObject<Selection>({
-			mode: params.mode,
-			list: params.list
-		});
+		if (params.projects) {
+			expect(params.query.projectSelection).toEqual(expect.anything());
+			expect(params.query.projectSelection).toMatchObject<Selection>({
+				mode: params.projects.mode,
+				list: params.projects.list
+			});
+		}
+		if (params.clients) {
+			expect(params.query.clientSelection).toEqual(expect.anything());
+			expect(params.query.clientSelection).toMatchObject<Selection>({
+				mode: params.clients.mode,
+				list: params.clients.list
+			});
+		}
 	} else {
 		expect(() => parser.parse(params.input, params.query)).toThrowError(
 			expectedError